Move selection details to separate file
