+++ src/blocks/right_sidebar.rs
@@ -1,44 +1,11 @@
 mod ontology_header;
+mod selection_details;
 
 use crate::{
-    blocks::right_sidebar::ontology_header::OntologyHeader,
-    components::{
-        accordion::Accordion, buttons::graph_interaction_buttons::GraphInteractionButtons,
-        user_input::internal_sparql::GraphDataContext,
-    },
-    events::EventContext,
+    blocks::right_sidebar::{ontology_header::OntologyHeader, selection_details::SelectionDetails},
+    components::buttons::graph_interaction_buttons::GraphInteractionButtons,
 };
 use leptos::prelude::*;
-
-#[component]
-pub fn SelectionDetails() -> impl IntoView {
-    let GraphDataContext { graph_metadata, .. } = expect_context::<GraphDataContext>();
-    let EventContext { show_metadata } = expect_context::<EventContext>();
-
-    let comments = create_read_slice(graph_metadata, |graph_metadata| {
-        graph_metadata.comments.clone()
-    });
-
-    let comment = Memo::new(move |_| {
-        if let Some(idx) = *show_metadata.read() {
-            comments
-                .read()
-                .get(&idx)
-                .cloned()
-                .unwrap_or_else(|| "No".to_string())
-        } else {
-            "No ".to_string()
-        }
-    });
-
-    view! {
-        <Accordion title="Selection Details">
-            <Show when=move || show_metadata.get().is_some() fallback=|| view! { <p>"Select an element in the visualization."</p>}>
-            <p>"Comment: "{move || comment.get()}</p>
-            </Show>
-        </Accordion>
-    }
-}
 
 #[component]
 pub fn RightSidebar() -> impl IntoView {
